requirements.txt
streamlit
psycopg2-binary
python-dotenv
requests
beautifulsoup4
python-dotenv>=0.19.0
